+++ app/src/main/java/com/ysshin/fine_dust_app/di/FragmentModule.kt
@@ -1,0 +1,10 @@
+package com.ysshin.fine_dust_app.di
+
+import com.ysshin.fine_dust_app.ui.LoginFragment
+import com.ysshin.fine_dust_app.ui.RegistrationFragment
+import org.koin.dsl.module
+
+val fragmentModule = module {
+    factory { LoginFragment() }
+    factory { RegistrationFragment() }
+}
