+++ app/src/main/java/com/ysshin/fine_dust_app/di/PreferenceModule.kt
@@ -1,0 +1,9 @@
+package com.ysshin.fine_dust_app.di
+
+import com.ysshin.fine_dust_app.data.PreferenceManager
+import org.koin.android.ext.koin.androidContext
+import org.koin.dsl.module
+
+val prefModule = module {
+    single { PreferenceManager(androidContext()) }
+}
