+++ app/src/main/java/com/ysshin/fine_dust_app/di/RepositoryModule.kt
@@ -1,0 +1,10 @@
+package com.ysshin.fine_dust_app.di
+
+import com.ysshin.fine_dust_app.data.LoginRepository
+import com.ysshin.fine_dust_app.data.RegistrationRepository
+import org.koin.dsl.module
+
+val repositoryModule = module {
+    single { LoginRepository(get()) }
+    single { RegistrationRepository(get()) }
+}
