+++ app/src/main/java/com/ysshin/fine_dust_app/di/ViewModelModule.kt
@@ -1,0 +1,11 @@
+package com.ysshin.fine_dust_app.di
+
+import com.ysshin.fine_dust_app.viewmodels.LoginViewModel
+import com.ysshin.fine_dust_app.viewmodels.RegistrationViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val viewModelModule = module {
+    viewModel { LoginViewModel(get()) }
+    viewModel { RegistrationViewModel(get()) }
+}
